requirements.txt

django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.10.1

gunicorn==20.0.4
keras==2.4.3
pandas==1.1.2
pillow==7.2.0

scikit-learn==0.23.2
tensorflow==2.3.1